style: Split values.rs into specified/computed module files
